Setting up my summing machine
Creating the HTML, CSS and JS files (that create the summing app).
App.js allows me to use Node and Express, to see the app on the localtunel:3000 and to share the app with the world.
PrintScreen and Recordings folders to show the app both on the localtunnel and on a functional website. The recordings also show the warning message if the user does not input any value.